+++ packages/pluggableWidgets/video-player-native/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   We have updated `react-native-video` version to 6.4.5
+
 ## [5.0.1] - 2024-8-29
 
 ### Fixed